feat: Introduce interactive map and compact views
Replaces the static `Locations` component with an interactive `InteractiveMapView` on the `ExplorePage`. This change also introduces a `CompactCountryList` and `CompactMapView` components, which are designed for more streamlined display of country data.

The `MapComponent` has been refactored to support dynamic fly-to functionality for future enhancements. Internationalization keys for these new components have also been added.